DAO and its implementation

+++ src/main/java/ru/octava/javarushtest/model/Book.java
@@ -5,7 +5,6 @@ import lombok.Data;
 import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
 import java.io.Serializable;
-import java.util.Date;
 
 @Entity
 @Table(name = "books")
@@ -14,7 +13,7 @@ public class Book implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @NotBlank
     private String title;
@@ -25,6 +24,7 @@ public class Book implements Serializable {
     private String author;
 
     private String isbn;
+
     private int print_year;
 
     @Column(nullable = false, columnDefinition = "TINYINT(4)", length = 1)
